backend/requirements.txt
fastapi
uvicorn
fastf1
anthropic
python-dotenv
pandas
numpy
